✨ Feat: 유저 정보 수정 api 연결 (#36)

+++ app/src/main/java/com/example/snack4diet/api/updateUser/UpdateUserDto.kt
@@ -2,8 +2,8 @@ package com.example.snack4diet.api.updateUser
 
 data class UpdateUserDto(
     val age: Int,
-    val height: Int,
+    val height: Double,
     val name: String,
-    val userId: Int,
-    val weight: Int
+    val userId: Long,
+    val weight: Double
 )
