+++ app/src/main/java/com/example/snack4diet/api/userInfo/Data.kt
@@ -2,7 +2,7 @@ package com.example.snack4diet.api.userInfo
 
 data class Data(
     val age: Int,
-    val height: Int,
+    val height: Double,
     val name: String,
-    val weight: Int
+    val weight: Double
 )
